# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UPAX Capital Humano Chatbot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface LoginRequest {
  employeeNumber: string;
  password: string;
}

export interface Employee {
  id: number;
  employeeNumber: string;
  name: string;
  businessUnit: string;
  role: string;
  hrbpName: string;
  hrbpPhoto: string;
  consultora?: string | null;
  isInternal: boolean;
  accentColor: string;
  logoUrl: string;
}

export interface LoginResponse {
  employee: Employee;
  message: string;
}

export interface ChatMessageRequest {
  message: string;
  category?: string | null;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  id: number;
  role: ChatMessageRole;
  content: string;
  timestamp: string;
  category?: string | null;
}

export interface ChatMessageResponse {
  message: ChatMessage;
  response: ChatMessage;
}

export interface ChatHistoryResponse {
  messages: ChatMessage[];
}