# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: UPAX Capital Humano Chatbot API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: chat
    description: Chat operations
  - name: employee
    description: Employee operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Employee login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Employee logout
      responses:
        "200":
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /auth/me:
    get:
      operationId: getCurrentEmployee
      tags: [auth]
      summary: Get current employee info
      responses:
        "200":
          description: Current employee
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Employee"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/message:
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a chat message and get AI response
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /chat/history:
    get:
      operationId: getChatHistory
      tags: [chat]
      summary: Get chat history for current employee
      responses:
        "200":
          description: Chat history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatHistoryResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    LoginRequest:
      type: object
      properties:
        employeeNumber:
          type: string
        password:
          type: string
      required:
        - employeeNumber
        - password

    Employee:
      type: object
      properties:
        id:
          type: integer
        employeeNumber:
          type: string
        name:
          type: string
        businessUnit:
          type: string
        role:
          type: string
        hrbpName:
          type: string
        hrbpPhoto:
          type: string
        consultora:
          type: string
          nullable: true
        isInternal:
          type: boolean
        accentColor:
          type: string
        logoUrl:
          type: string
      required:
        - id
        - employeeNumber
        - name
        - businessUnit
        - role
        - hrbpName
        - hrbpPhoto
        - isInternal
        - accentColor
        - logoUrl

    LoginResponse:
      type: object
      properties:
        employee:
          $ref: "#/components/schemas/Employee"
        message:
          type: string
      required:
        - employee
        - message

    ChatMessageRequest:
      type: object
      properties:
        message:
          type: string
        category:
          type: string
          nullable: true
      required:
        - message

    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        timestamp:
          type: string
        category:
          type: string
          nullable: true
      required:
        - id
        - role
        - content
        - timestamp

    ChatMessageResponse:
      type: object
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        response:
          $ref: "#/components/schemas/ChatMessage"
      required:
        - message
        - response

    ChatHistoryResponse:
      type: object
      properties:
        messages:
          type: array
          items:
            $ref: "#/components/schemas/ChatMessage"
      required:
        - messages