# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UPAX Capital Humano Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Employee login
 */
export const LoginBody = zod.object({
  employeeNumber: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  employee: zod.object({
    id: zod.number(),
    employeeNumber: zod.string(),
    name: zod.string(),
    businessUnit: zod.string(),
    role: zod.string(),
    hrbpName: zod.string(),
    hrbpPhoto: zod.string(),
    consultora: zod.string().nullish(),
    isInternal: zod.boolean(),
    accentColor: zod.string(),
    logoUrl: zod.string(),
  }),
  message: zod.string(),
});

/**
 * @summary Employee logout
 */
export const LogoutResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get current employee info
 */
export const GetCurrentEmployeeResponse = zod.object({
  id: zod.number(),
  employeeNumber: zod.string(),
  name: zod.string(),
  businessUnit: zod.string(),
  role: zod.string(),
  hrbpName: zod.string(),
  hrbpPhoto: zod.string(),
  consultora: zod.string().nullish(),
  isInternal: zod.boolean(),
  accentColor: zod.string(),
  logoUrl: zod.string(),
});

/**
 * @summary Send a chat message and get AI response
 */
export const SendChatMessageBody = zod.object({
  message: zod.string(),
  category: zod.string().nullish(),
});

export const SendChatMessageResponse = zod.object({
  message: zod.object({
    id: zod.number(),
    role: zod.enum(["user", "assistant"]),
    content: zod.string(),
    timestamp: zod.string(),
    category: zod.string().nullish(),
  }),
  response: zod.object({
    id: zod.number(),
    role: zod.enum(["user", "assistant"]),
    content: zod.string(),
    timestamp: zod.string(),
    category: zod.string().nullish(),
  }),
});

/**
 * @summary Get chat history for current employee
 */
export const GetChatHistoryResponse = zod.object({
  messages: zod.array(
    zod.object({
      id: zod.number(),
      role: zod.enum(["user", "assistant"]),
      content: zod.string(),
      timestamp: zod.string(),
      category: zod.string().nullish(),
    }),
  ),
});