# lib/api-zod/src/generated/types/chatMessageResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UPAX Capital Humano Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { ChatMessage } from "./chatMessage";

export interface ChatMessageResponse {
  message: ChatMessage;
  response: ChatMessage;
}