# lib/api-zod/src/generated/types/employee.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UPAX Capital Humano Chatbot API
 * OpenAPI spec version: 0.1.0
 */

export interface Employee {
  id: number;
  employeeNumber: string;
  name: string;
  businessUnit: string;
  role: string;
  hrbpName: string;
  hrbpPhoto: string;
  consultora?: string | null;
  isInternal: boolean;
  accentColor: string;
  logoUrl: string;
}