# lib/api-zod/src/generated/types/loginResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * UPAX Capital Humano Chatbot API
 * OpenAPI spec version: 0.1.0
 */
import type { Employee } from "./employee";

export interface LoginResponse {
  employee: Employee;
  message: string;
}